unified train/inference
